errors fixed

+++ src/alg.cpp
@@ -6,7 +6,7 @@
 
 bool checkPrime(uint64_t value) {
   if (value < 2) return false;
-  for (size_t i = 2; i * i <= value; i++) {
+  for (int i = 2; i * i <= value; i++) {
     if (value % i == 0) return false;
   }
   return true;
@@ -14,14 +14,14 @@ bool checkPrime(uint64_t value) {
 
 uint64_t nPrime(uint64_t n) {
   int num = 1;
-  for (size_t i = 2; i <= n; i++) {
+  for (int i = 2; i <= n; i++) {
     if (checkPrime(i) == true) num++;
   }
   return num;
 }
 
 uint64_t nextPrime(uint64_t value) {
-  for (size_t i = value + 1; i < (value * value); i++) {
+  for (int i = value + 1; i < (value * value); i++) {
     if (checkPrime(i) == true) return i;
   }
   return 2;
@@ -39,7 +39,7 @@ uint64_t sumPrime(uint64_t hbound) {
 
 uint64_t twinPrimes(uint64_t lbound, uint64_t hbound) {
   int count = 0;
-  for (size_t i = lbound; i < hbound; i++) {
+  for (int i = lbound; i < hbound; i++) {
     if ((checkPrime(i) == true) && (checkPrime(i+2) == true)) count ++;
   }
   return count;
